% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{add}
\alias{add}
\alias{subtract}
\alias{multiply}
\alias{divide}
\alias{xtract}
\alias{xxtract}
\alias{dollar}
\title{Convinient for R4 pipe}
\usage{
add()

subtract()

multiply()

divide()

xtract()

xxtract()

dollar()
}
\description{
adding two numbers
}
