% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{adj_to_precision}
\alias{adj_to_precision}
\title{Generate precision matrix for spatial model from matrix of adjacency}
\usage{
adj_to_precision(x)
}
\arguments{
\item{x}{matrix of connectivity with 1s denote connected nodes and zeros otherwise.}
}
\description{
Generate precision matrix for spatial model from matrix of adjacency
}
