% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{allot}
\alias{allot}
\title{Right assign to using with pipe}
\usage{
allot(.data, name)
}
\arguments{
\item{.data}{data to assign}

\item{name}{name of object to assign}
}
\description{
Using \code{->} with pipe
}
\examples{
x <- 1
allot(2, x)
3 \%>\% allot(x)
x
}
