% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{AR2_Q}
\alias{AR2_Q}
\title{Autoregressive order 2 - AR(2) precision matrix generator}
\usage{
AR2_Q(n = 10, sd = 1, rho = c(0.9, 0.05))
}
\arguments{
\item{n}{length}

\item{sd}{standard deviation}

\item{rho}{two AR2 coefficients}
}
\description{
Autoregressive order 2 - AR(2) precision matrix generator
}
\examples{
QQ <- AR2_Q(10)
x <- INLA::inla.qsample(1, Q, 
  constr = list(A = matrix(1, ncol = nrow(Q)), e = 0))
plot(x, type = 'l')
}
