% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\name{as_field}
\alias{as_field}
\title{Convert a R list of variable to mongo fields}
\usage{
as_field(...)
}
\arguments{
\item{...}{write a list of varible to get}
}
\description{
To filter the data (select columns)
}
\examples{
\dontrun{
as_field(var1, var2)
}
}
