% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\name{as_or}
\alias{as_or}
\title{Convert an R named list to mongo `OR`}
\usage{
as_or(...)
}
\arguments{
\item{...}{a named list}
}
\description{
Convert an R named list to mongo `OR`
}
\examples{
\dontrun{
query = as_or(interview = "Main interview", interview = "COVID-19 CATI interview (ca)")
}
}
