% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\name{as_query}
\alias{as_query}
\title{Convert a R named list of variable and condition to mongo query}
\usage{
as_query(...)
}
\arguments{
\item{...}{write a list of varible to query}
}
\description{
To fillter the data (select rows)
}
\examples{
\dontrun{
# not there is only one `=`
as_query(interview = TRUE)
}
}
