% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_utils.R
\name{cd}
\alias{cd}
\title{replicate some features of bash cd}
\usage{
cd(to = "~")
}
\arguments{
\item{x}{x}
}
\description{
replicate some features of bash cd such as go backs to previous working directory with cd('-')
}
