% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{char}
\alias{char}
\title{Convenient create vector of character without the need to quote}
\usage{
char(...)
}
\arguments{
\item{...}{list of character separate by comma}
}
\description{
Convenient create vector of character without the need to quote, e.g. instead
of typing \code{c("a", "b", "c")} just write \code{char(a, b, c)}
}
