% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{claim}
\alias{claim}
\title{Claim columns from a data frame and store them globally}
\usage{
claim(df, ..., .name = ".kollection")
}
\arguments{
\item{df}{The input data frame from which to claim columns.}

\item{...}{Unquoted column names from `df` to be claimed.}

\item{.name}{The name of the global data frame to store the claimed columns.
Defaults to `".kollection"`.}
}
\value{
Invisibly returns the original data frame `df`.
}
\description{
This function allows you to "claim" specified columns from a data frame
and store them in a global data frame named `.kollection` (or a custom name).
If `.kollection` doesn't exist, it's created. If it exists but has a different
number of rows, it's reinitialized to match the input data frame's row count.
}
