% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{corfrankCopula}
\alias{corfrankCopula}
\title{Frank copulas parameter to kendall tau or spearman rho}
\usage{
corfrankCopula(x, type = c("spearman", "kendall"))
}
\description{
Frank copulas parameter to kendall tau or spearman rho
}
\examples{
corfrankCopula(1, 'spearman')
}
