% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_lines.R
\name{count_lines}
\alias{count_lines}
\title{Code Stats}
\usage{
count_lines(con, verbose = FALSE)
}
\arguments{
\item{con}{a connection object or a character string (passed to \link[base]{readLines}).}

\item{verbose}{a logical (FALSE by default) if counts should be printed to the console.}
}
\value{
a named vector.
}
\description{
Builds a basic code counts report.
}
\details{
The report will contain the following information:
\itemize{
\item total
\item code
\item comment
\item documentation
\item spacing
}

Code lines ending with a comment are counted as code lines (hence ignored from comments).
}
\examples{
infile <- tempfile()

lines <- "#' This\n#'\n\n# comment\np <- 1\n\nif(p==1) print(p)\n"

writeLines(lines, infile)
count_lines(infile)
}
\seealso{
\code{\link[=scan_code]{scan_code()}}
}
