% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_paitwise.R
\name{count_pairwise}
\alias{count_pairwise}
\title{Count pairs of items within a group}
\usage{
count_pairwise(tbl, item, feature, sort = FALSE, diag = FALSE, upper = TRUE)
}
\arguments{
\item{tbl}{Table}

\item{item}{Item to count pairs of; will end up in \code{item1} and
\code{item2} columns}

\item{feature}{Column within which to count pairs
\code{item2} columns}

\item{sort}{Sort the output by \code{n}?}

\item{diag}{Count.}

\item{upper}{Remove duplicates.}
}
\description{
Count the number of times each pair of items appear together within a group
defined by "feature." For example, this could count the number of times
two words appear within documents).
}
\examples{
\dontrun{
library(dplyr)
dat <- tibble(group = rep(1:5, each = 2),
              letter = c("a", "b",
                         "a", "c",
                         "a", "c",
                         "b", "e",
                         "b", "f"))

# count the number of times two letters appear together
count_pairwise(dat, letter, group)
count_pairwise(dat, letter, group, sort = TRUE)
count_pairwise(dat, letter, group, sort = TRUE, diag = TRUE)
}

}
