% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data}
\alias{create_data}
\title{Create Data}
\usage{
create_data(colClasses)
}
\arguments{
\item{colClasses}{a named vector of classes.}
}
\value{
a data.frame.
}
\description{
Build an empty data.frame from a vector of colClasses.
}
\details{
The names of the vector will be used to name the columns and the values
to cast the columns to the given classes.

If \code{NA} is given as input for colClasses, then read.table() will return
an empty data.frame (0 obs. of 0 variables)
}
\examples{
create_data(colClasses = c(id = "numeric", date = "Date", comment = "character"))
}
