% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dgumbelCopula}
\alias{dgumbelCopula}
\title{gumbel copulas density}
\usage{
dgumbelCopula(u1, u2, a, give_log = FALSE)
}
\arguments{
\item{u1}{dimension 1}

\item{u2}{dimension 2}

\item{alpha}{copula parameter}
}
\description{
gumbel copulas density
}
