% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{double_logistic}
\alias{double_logistic}
\title{Double logistic function}
\usage{
double_logistic(
  x = seq(0, 10, 0.01),
  bound1 = 0,
  bound2 = 1,
  rate1 = 2,
  rate2 = 5,
  midpoint1 = 3,
  midpoint2 = 7
)
}
\arguments{
\item{x}{time horizon}

\item{bound1}{the boundary of function (this will equals the function value at time zero)}

\item{bound2}{the second boundary, if bound1 > bound2 the function increase first then decrease, and vice versa.}

\item{rate1}{rate of change of first period, control the sharpness of changes}

\item{rate2}{rate of change of second period, control the sharpness of changes}

\item{midpoint1}{midpoint where the chance occurs in the first period}

\item{midpoint2}{midpoint where the chance occurs in the second period}
}
\value{
vector of length x
}
\description{
Double logistic function  have two periods each with a logistic shape, either one of them will be decreasing and the other increases.
}
\examples{
plotl(double_logistic(bound1=0.1, bound2=0.8))
plotl(double_logistic(bound1=0.8, bound2=0.1))

}
\seealso{
\code{\link{logistic}}
}
