% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{download}
\alias{download}
\title{Download with httr}
\usage{
download(url, path = ".", overwrite = FALSE, ...)
}
\arguments{
\item{url}{directly link}

\item{path}{where to save}
}
\description{
wrapper with automatic nameing, urldecode, and progress
}
