% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{dttb}
\alias{dttb}
\title{Display HTML with common options}
\usage{
dttb(x, ...)
}
\arguments{
\item{x}{a data frame}

\item{...}{options to DT::datatable}
}
\description{
Display HTML with common options
}
