% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Emojis}
\alias{Emojis}
\title{All Unicode-Emojis}
\format{A data frame with 1719 rows and 4 variables:
\describe{
  \item{Emoji}{Emoji character}
  \item{cldr_short_name}{Unicode Common Locale Data Repository (CLDR) short name}
  \item{Code}{UTF-8 codepoint}
  \item{short_name}{A short name containing no spaces, based on the CLDR short name}
}}
\usage{
Emojis
}
\description{
A dataset containing the Emoji-characters, codepoints, cldr-short names and descriptions for all Emojis.
}
\keyword{datasets}
