% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{F_gllogisI}
\alias{F_gllogisI}
\title{Generalized log-logistic type I probability density}
\usage{
F_gllogisI(t, lambda, p, gamma)
}
\description{
Generalized log-logistic type I probability density
}
