% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{f_glogisI}
\alias{f_glogisI}
\alias{F_glogisI}
\title{Generalized logistic type I density}
\usage{
f_glogisI(x, alpha, beta, gamma)

F_glogisI(x, alpha, beta, gamma)
}
\arguments{
\item{x}{x}

\item{alpha}{scale}

\item{beta}{shape}

\item{gamma}{skewness}
}
\description{
Generalized logistic type I density

Generalized logistic type I cumulative
}
