% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{file_name}
\alias{file_name}
\title{clean file name}
\usage{
file_name(path, keep_extension = FALSE, escape_space = FALSE, ...)
}
\arguments{
\item{path}{string of path/url}

\item{keep_extension}{or not?}

\item{escape_space}{or not?, escape space in file name}
}
\description{
clean file name
}
