% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_utils.R
\name{fill_polygon_h3}
\alias{fill_polygon_h3}
\title{Fill `sf` with `h3` indexes}
\usage{
fill_polygon_h3(.data, res = 7, ts = 4326)
}
\arguments{
\item{res}{resolution, 7 = 5km2, 6 = ~36km2, ...}

\item{ts}{for st_transform if needed}

\item{df}{`sf` data}
}
\value{
a vector of h3 indexes of all the polygons in the supplied data
}
\description{
Fill `sf` with `h3` indexes
}
