% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_date.R
\name{filter_date}
\alias{filter_date}
\title{Filter by Date Range}
\usage{
filter_date(x, colname = NULL, ref = Sys.Date(), unit = "month")
}
\arguments{
\item{x}{a data.frame with a Date or POSIXct column.}

\item{colname}{an optional character string, the name of the column to use in filter.}

\item{ref}{an optional Date value to be used as a reference for the date range.}

\item{unit}{a string to compute the date range (default = "month") (see \link[lubridate]{round_date}).}
}
\value{
a filtered data.frame.
}
\description{
Filter a data.frame by specific date range.
}
\details{
The purpose of this function is to make data navigation between date range easier.
By default, it will filter the input data.frame to keep dates belonging to the current month.

When \code{colname = NULL} (the default), it will look for a 'date' column in x.
}
\examples{
# -- this month (the default)
filter_date(data.frame(date = Sys.Date() - runif(n = 10, min = -50, max = 50)))

# -- previous month
filter_date(data.frame(date = Sys.Date() - runif(n = 10, min = -50, max = 50)),
ref = Sys.Date() - as.integer(format(Sys.Date(), "\%d")))

}
