% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{find_in_zip}
\alias{find_in_zip}
\title{find filename matching regex in a zipped file}
\usage{
find_in_zip(x, pattern, basename = FALSE, ...)
}
\arguments{
\item{x}{character zip file path}

\item{pattern}{character regex pattern}

\item{basename}{boolean find in base name only (exclude path but with extensiion)}

\item{...}{extra regex's args}
}
\description{
find filename matching regex in a zipped file
}
