% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_utils.R
\name{findInterval2}
\alias{findInterval2}
\title{findInterval and return factor with label}
\usage{
findInterval2(x, vec, factor = TRUE, ...)
}
\description{
instead of numeric vector like the base version
}
