% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{format_pvalue}
\alias{format_pvalue}
\title{format p-value for publishing}
\usage{
format_pvalue(x, inequality = TRUE, digits = 2)
}
\arguments{
\item{x}{computed p-value}
}
\description{
format p-value for publishing
}
