% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly.R
\name{fractional_poly}
\alias{fractional_poly}
\title{Find best fit fractional polynomial using GLM}
\usage{
fractional_poly(x, y, verbose = FALSE, ...)
}
\arguments{
\item{x}{indepedent variable to find best fit fractional powers}

\item{y}{dependent variable, no transformation is done in the code}

\item{...}{arguments to glm()}
}
\description{
Find best fit fractional polynomial using GLM
}
