% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{gen_widths}
\alias{gen_widths}
\title{Generate widths for geom_tile with unevenly space x-axis}
\usage{
gen_widths(x)
}
\arguments{
\item{x}{x-axis vector to plot}
}
\value{
weights use in \code{\link[ggplot2]{geom_tile}}
}
\description{
Calculate the widths needed to fill the white space when we have unevenly spaced x-axis value
}
\examples{
\dontrun{
df <- data.frame(
  x = rep(c(2, 5, 7, 9, 12), 2),
  y = rep(c(1, 2), each = 5),
  z = factor(rep(1:5, each = 2))
)
w = gen_widths(df$x)
ggplot(df, aes(x, y)) +geom_tile(aes(fill = z, width=w), colour = "white")
}
}
