% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{get_report}
\alias{get_report}
\title{Get values of summary(\code{\link[TMB]{sd_report}}) by name}
\usage{
get_report(x, name = ".*", se = TRUE)
}
\arguments{
\item{x}{object}

\item{name}{regex of parameter's name}

\item{se}{get the standard error as well}
}
\description{
Get values of summary(\code{\link[TMB]{sd_report}}) by name
}
