% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{getcoding_readstata13}
\alias{getcoding_readstata13}
\title{Read variable coding of a variable from stata file}
\usage{
getcoding_readstata13(data, var_name)
}
\arguments{
\item{data}{data}

\item{var_name}{character name of variable}
}
\description{
Read variable coding of a variable from stata file
}
