% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{getmeta_readstata13}
\alias{getmeta_readstata13}
\title{Read meta data from stata file}
\usage{
getmeta_readstata13(data)
}
\arguments{
\item{data}{data}
}
\description{
This includes labels, coding values
}
