% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNsFunction.R
\name{getNsFunction}
\alias{getNsFunction}
\title{Package Function in Do Call}
\usage{
getNsFunction(x)
}
\arguments{
\item{x}{the name of the function, most probably as pkg::fun()}
}
\value{
the exported function name in namespace pkg
}
\description{
Trick to solve the use of :: for package functions in do.call()
}
\examples{
\dontrun{
getNsFunction("pkg::fun")
}
}
