% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimestamp.R
\name{getTimestamp}
\alias{getTimestamp}
\title{Compute Timestamp}
\usage{
getTimestamp(k = 1000, digits = 0, silent = FALSE)
}
\arguments{
\item{k}{a numeric (default = 1000), used as a multiplication factor.}

\item{digits}{an integer (default = 0) to round the value.}

\item{silent}{an optional (default = FALSE) logical. If TRUE, no traces will go to the console.}
}
\value{
a numeric.
}
\description{
Compute a numeric timestamp.
}
\details{
By default (\code{k = 1000}), the function returns a numeric up to the millisecond.

Output should not be used as a unique id if users / systems may call it more than
one time per millisecond.
}
\examples{
# compute up to the second timestamp
getTimestamp(k = 1, digits = 0)

# compute up to the millisecond timestamp
getTimestamp(k = 1000)
getTimestamp()

}
\seealso{
\code{\link[=uuid]{uuid()}}
}
