% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gg_group}
\alias{gg_group}
\title{Prepare ggplot by groups}
\usage{
gg_group(.data, x, y, g)
}
\arguments{
\item{.data}{data contains variables}

\item{x}{x-axis}

\item{y}{y-axis}

\item{g}{grouping variabe}
}
\value{

}
\description{
Prepare ggplot by groups
}
