% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggcol}
\alias{ggcol}
\title{Default column plot for ktools}
\usage{
ggcol(.data, x, y, p = position_dodge())
}
\arguments{
\item{.data}{data}

\item{x}{aes x}

\item{y}{aes y}

\item{p}{position **function**}
}
\description{
Default column plot for ktools
}
