% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gghist}
\alias{gghist}
\title{Default histogram plot for ktools}
\usage{
gghist(.data, x, bins = 30)
}
\arguments{
\item{.data}{data}

\item{x}{aes x}

\item{bins}{bins for histogram}
}
\description{
Default histogram plot for ktools
}
