% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_utils.R
\name{grept}
\alias{grept}
\title{Find text match regex with grep with some tailored options}
\usage{
grept(pattern, x, ignore.case = 1, useBytes = TRUE, value = TRUE, ...)
}
\arguments{
\item{...}{grep's args}
}
\description{
Overidding is OK
}
