% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_utils.R
\name{hex_fill}
\alias{hex_fill}
\title{Find most frequent occurrences in the neighbors and fill in the missing
hexagon}
\usage{
hex_fill(
  .data,
  ids,
  var,
  h3var = "h3",
  radius_max = 3,
  mode = c("same", "progressive")
)
}
\arguments{
\item{.data}{data with h3 columm}

\item{ids}{h3 ids to find}

\item{var}{variable to fill}

\item{h3var}{name of the h3 column}

\item{radius_max}{maximum of radius of hexagons to find}

\item{mode}{use the same number of neighbours (same) or only increase neighbors
in those hexagon with missing data at the smaller radius (progressive)}
}
\description{
Find most frequent occurrences in the neighbors and fill in the missing
hexagon
}
\details{
progressive maximum the local influences while the other treats all missing
hexagon the same.
}
