% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.unique.R
\name{is.unique}
\alias{is.unique}
\title{Does a list contain duplicates.}
\usage{
is.unique(list, boolOnly = TRUE)
}
\arguments{
\item{list}{The list to check for uniqueness.}

\item{boolOnly}{Return only a boolean.}
}
\value{
If \code{boolOnly == FALSE} then only \code{TRUE}/\code{FALSE} is returned, else print table with length of \code{list} and \code{unique(list)}.
}
\description{
Check if \code{length(x) == length(unique(x))} and optionally print table.
}
\examples{
\dontrun{
is.unique(list)
# [1] TRUE

is.unique(x, boolOnly = F)
# complete   unique    delta
# 11       10        1
# [1] FALSE
}

}
