% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_utils.R
\name{join_list}
\alias{join_list}
\title{Join elements of list a from list b}
\usage{
join_list(of_a, from_b, add_new = TRUE, replace = TRUE)
}
\arguments{
\item{of_a}{list to replace}

\item{from_b}{list with elements to take}

\item{add_new}{add new elements if not existed}

\item{replace}{TRUE replace elements with the same name}
}
\description{
Join elements of list a from list b
}
