% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdebug.R
\name{k_debug}
\alias{k_debug}
\title{Set Trace Level}
\usage{
k_debug(level = NULL)
}
\arguments{
\item{level}{a numeric vector to indicate the level of trace printed by
\link[ktools]{catl} function or NULL to set the trace OFF}
}
\description{
Set Trace Level
}
\examples{
k_debug()
k_debug(2)
}
