% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{KLD}
\alias{KLD}
\title{Kullback Liebler}
\usage{
KLD(target, estimate)
}
\arguments{
\item{target}{target distribution density}

\item{estimate}{estimate distribution density that we wish to compare}
}
\description{
Kullback Liebler
}
