% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{kompile}
\alias{kompile}
\title{Compile TMB with ktools header}
\usage{
kompile(..., user_flags, verbose = FALSE)
}
\arguments{
\item{...}{Passed as Makeconf variables.}

\item{user_flags}{extra flag to compile}

\item{verbose}{print}
}
\description{
Add ktools.hpp to compile flags
}
