% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_utils.R
\name{kring_smooth}
\alias{kring_smooth}
\title{K-ring smoothing}
\usage{
kring_smooth(df, hex, metric, k = 1, only_na = TRUE)
}
\arguments{
\item{df}{data with h3 column}

\item{hex}{character name of the h3 column}

\item{metric}{character name of the metric column (continous variable)}

\item{k}{ring's radius}

\item{only_na}{fill missing values only, otherwise smooth all}
}
\description{
K-ring smoothing
}
