% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{kut}
\alias{kut}
\title{Cut but automatically include min and max data's value}
\usage{
kut(x, breaks, ...)
}
\arguments{
\item{x}{a numeric vector which is to be converted to a factor by cutting.}

\item{breaks}{either a numeric vector of two or more unique cut points or a
    single number (greater than or equal to 2) giving the number of
    intervals into which \code{x} is to be cut.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Cut but automatically include min and max data's value
}
