% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{label_table}
\alias{label_table}
\title{Generate lookup table for haven read spss/stata}
\usage{
label_table(d)
}
\arguments{
\item{d}{data}
}
\description{
Generate lookup table for haven read spss/stata
}
