% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{labelOutlier}
\alias{labelOutlier}
\title{labelOutlier}
\usage{
labelOutlier(x, y = NULL, alpha = 0.01, at = 3, n = 5, ...)
}
\description{
Add labels (id) for data points with farthest \code{mahalanobis} distance.
}
