% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_pls.R
\name{library_pls}
\alias{library_pls}
\title{Library please! Easier Package Installation}
\usage{
library_pls(package)
}
\arguments{
\item{package}{The package to install.}
}
\description{
If the last error was raised due to calling \code{library()} on a missing package.
}
\examples{
\dontrun{
library(parallel)
# Error in library(parallel) : there is no package called ‘parallel’

library_pls(parallel)
# Install package parallel?
# (y/n) y
# Installing package …
}

}
