% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{list_name_to_column}
\alias{list_name_to_column}
\title{Add nested list name to nested data column}
\usage{
list_name_to_column(a_list, col_name = "ele_name")
}
\arguments{
\item{a_list}{list contains nested data frame}
}
\description{
Add nested list name to nested data column
}
