% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{list_zip}
\alias{list_zip}
\title{List zip file contents with unzip or unar (default)}
\usage{
list_zip(
  zipfile,
  long = FALSE,
  verylong = FALSE,
  password,
  encoding,
  print_encoding = FALSE,
  indexes = FALSE,
  ...
)
}
\arguments{
\item{zipfile}{zip file}

\item{long}{Print more information about each file in the archive.}

\item{verylong}{Print all available information about each file in the archive.}

\item{password}{The password to use for decrypting protected archives.}

\item{encoding}{The encoding to use for filenames in the archive, when it is not known. If not specified, the program attempts to auto-detect the encoding used.}

\item{print_encoding}{Print the auto-detected encoding and the confidence factor after the file list}

\item{indexes}{Instead of specifying the files to list as filenames or wildcard patterns, specify them as indexes.}
}
\description{
Just to easy to remember the function name and alternate between unar and unzip
}
