% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{load_pkg}
\alias{load_pkg}
\title{Wrapper of suppress message library loadings}
\usage{
load_pkg(...)
}
\arguments{
\item{...}{name of package that is normally loaded with `library`}
}
\description{
Wrapper of suppress message library loadings
}
