% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{log_prec_prior}
\alias{log_prec_prior}
\title{Log penalized complexity precision prior like INLA's internal}
\usage{
log_prec_prior(log_prec, u, alpha)
}
\arguments{
\item{log_prec}{log precision}

\item{u}{u}

\item{alpha}{alpha}
}
\description{
Log penalized complexity precision prior like INLA's internal
}
\examples{
sd <- seq(0.001, 2, 0.01)
prec <- sd2prec(sd)
lprec <- log(prec)
log_prec_prior(lprec, 0.33, 0.01) \%>\%
  exp() \%>\%
  plot(sd, .)
log_prec_prior(lprec, 1, 0.01) \%>\%
  exp() \%>\%
  lines(sd, .)
}
