% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{logistic}
\alias{logistic}
\title{Logistic shape generator}
\usage{
logistic(
  time = seq(0, 1, 0.01),
  initial = 1,
  percent_reduction = 0.5,
  midpoint = 0.5,
  shape = 1
)
}
\arguments{
\item{time}{time horizon}

\item{initial}{value at time zero}

\item{percent_reduction}{in [0, 1], percentage reduction compare to the initial value after midpoint of the time horizon, so the end value equals initial x percent_reduction}

\item{midpoint}{the time the function changes}

\item{shape}{control how sharp the transition from initial value to the end value}
}
\description{
Logistic shape generator
}
