% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{long_matrix}
\alias{long_matrix}
\title{Matrix with names to long table with rowname as column}
\usage{
long_matrix(m)
}
\arguments{
\item{m}{matrix with rowname}
}
\description{
Matrix with names to long table with rowname as column
}
