% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.option.R
\name{match.option}
\alias{match.option}
\title{Match Options}
\usage{
match.option(fun, arg, value)
}
\arguments{
\item{fun}{the reference of the function, without ().}

\item{arg}{the name of the argument to test.}

\item{value}{the value to test.}
}
\value{
The list of options, including missing ones.
}
\description{
Matches a named list against the one expected in function arguments.
}
\examples{

# function expecting a list as argument
foo <- function(options = list(a = 1, b = 2)){}

# match user input
match.option(foo, "options", list(a = 1))
}
