% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{meanstat}
\alias{meanstat}
\title{format cross table for publushing}
\usage{
meanstat(.data, y, g, method = c("t.test", "wilcox.test"))
}
\arguments{
\item{y}{continouus/count}

\item{g}{grouping variable}
}
\description{
t-test and wilcoxon
}
