% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{median_gllogisI}
\alias{median_gllogisI}
\title{Generalized log-logistic type I probability density}
\usage{
median_gllogisI(lambda, p, gamma)
}
\description{
Generalized log-logistic type I probability density
}
