% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modus.R
\name{modus}
\alias{modus}
\title{Calculate mode.}
\usage{
modus(x, na.rm = FALSE)
}
\arguments{
\item{x}{An R object. Currently there are methods for numeric objects.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
The mode of \code{x} or a list of the modes of \code{x}.
}
\description{
Generic function for mode of a sample.
}
\examples{
modus(c(1,1,1,2,2,3,NA), na.rm = TRUE)

}
