% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mu_skew_llogis}
\alias{mu_skew_llogis}
\title{Compute the mean of skew loglogistic distribution}
\usage{
mu_skew_llogis(scale, shape, skew)
}
\arguments{
\item{scale}{scale}

\item{shape}{shape}

\item{skew}{skewness parameters}
}
\description{
Compute the mean of skew loglogistic distribution
}
