% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{mustats_3plus}
\alias{mustats_3plus}
\title{format cross-IQR by groups (>2) for publishing}
\usage{
mustats_3plus(.data, ...)
}
\arguments{
\item{.data}{pipe or data}

\item{...}{table's args}
}
\description{
format cross-table for publishing, only kruskal wallis/bonferonni atm
}
