% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{name2iso}
\alias{name2iso}
\title{shorthand country name to ISO from countrycode package with some custom match}
\usage{
name2iso(x, version = 3, ...)
}
\description{
shorthand country name to ISO from countrycode package with some custom match
}
