% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{name2list}
\alias{name2list}
\title{Convert named vector to a named list}
\usage{
name2list(x)
}
\arguments{
\item{x}{a named vector}
}
\description{
Main usage is extracting named parameters returned by e.g., \code{\link[base]{optim}}
}
\examples{
x <- c(a = 1, b = 2)
name2list(x)
}
