% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\docType{data}
\name{name_collection}
\alias{name_collection}
\alias{otherwise}
\alias{is_missing}
\title{Collection of naming, renaming functions, values}
\format{
An object of class \code{logical} of length 1.
}
\usage{
otherwise

is_missing()
}
\description{
Collection of naming, renaming functions, values

name TRUE as `otherwise`

name `is_missing` = `is.na`
}
\keyword{datasets}
