% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{napply}
\alias{napply}
\title{n(amed)apply: lapply but automatic add names to output}
\usage{
napply(x, ...)
}
\arguments{
\item{x}{lappy args}

\item{...}{lappy args}
}
\description{
lapply but automatic add names to output
}
