% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{network_to_precision}
\alias{network_to_precision}
\title{Generate precision matrix for spatial model from matrix of network}
\usage{
network_to_precision(x)
}
\arguments{
\item{x}{matrix of connectivity with 1s denote connected nodes and zeros otherwise.}
}
\description{
Generate precision matrix for spatial model from matrix of network
}
