% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{nullspace_penalty}
\alias{nullspace_penalty}
\title{Null-space penalty of the precision matrix}
\usage{
nullspace_penalty(x)
}
\arguments{
\item{x}{a precision matrix}
}
\description{
This add e.g., constraints zero intercept and slope for second-order random
walk model.
}
\details{
Using eigen decomposition to find the zero eigenvalues, which is then added
back to the original penalized matrix. Note that this leads to loss of
sparseness.
}
