% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_adj.R
\name{p_adj}
\alias{p_adj}
\title{Adjusted prevalence using sensitivity and specificity}
\usage{
p_adj(n, x, alpha = 3, beta = 15, sens = 0.995, spec = 0.9)
}
\arguments{
\item{n}{sample size}

\item{x}{number of cases}

\item{alpha}{shape1 of beta distribution prior}

\item{beta}{shape2 of beta distribution prior}

\item{sens}{sensitivity of the test}

\item{spec}{specificity of the test}
}
\description{
Using a modified Lew and Levy  to allow non-uniform prior on the prevalenc
}
