% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_table.R
\name{p_table}
\alias{p_table}
\title{Faster \code{prop.table(table())}}
\usage{
p_table(object, percent = TRUE, ...)
}
\arguments{
\item{object}{An object to tabulate.}

\item{percent}{Return rounded percentage.}

\item{...}{Other arguments passed to \link[base]{table}.}
}
\value{
\code{p_table()} returns a contingency table, an object of class "table", an array of integer values. See \link[base]{table}.
}
\description{
Wrapper for \code{table()} and \code{prop.table()}. Returns APA compliant percentage values from table data.
}
\examples{
\dontrun{
p_table(df, percent = FALSE, useNA = 'always')
}

}
