% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{panel.smooth}
\alias{panel.smooth}
\title{panel.smooth}
\usage{
panel.smooth(
  x,
  y,
  col = par("col"),
  bg = NA,
  pch = par("pch"),
  cex = 1,
  col.smooth = "red",
  span = 2/3,
  iter = 3,
  ...
)
}
\description{
Panel for plotting pair()
}
