% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteNaRm.R
\name{pasteNArm}
\alias{pasteNArm}
\title{Alternative to `paste()`, where NAs are remove before conversion to character strings}
\usage{
pasteNArm(..., sep = "", collapse = "")
}
\arguments{
\item{...}{See `paste()` for arguments.}

\item{sep}{See `paste()` for arguments.}

\item{collapse}{See `paste()` for arguments.}
}
\value{
See `paste()` for return values.
}
\description{
Behaves as `paste()` would, but removes NAs before converting its arguments (via as.character) to character strings
}
\examples{
\dontrun{
s <- c("a", "b", NA, "d")
paste(s, collapse = "")
# [1] "abNAd"

pasteNArm(s, collapse = "")
# [1] "Abd"
}

}
