% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcount.R
\name{pcount}
\alias{pcount}
\title{Wrapper for \code{count()} with percentages}
\usage{
pcount(tbl, x, rounded = TRUE, name = "n", ...)
}
\arguments{
\item{tbl}{Table}

\item{x}{The variable to count and calculate the percentage on.}

\item{rounded}{logical. Return rounded percentage. If false decimal values will be used.}

\item{name}{character. The name of the new column in the output. If omitted, it will default to n.}

\item{...}{Other arguments passed to \link[dplyr]{count}.}
}
\value{
\code{pcount()} returns the same as \code{count()} with an aditional percentage column. See \link[dplyr]{count}.
}
\description{
Shortcut for count(x) %>% mutate(p = n/sum(n)). Returns APA compliant percentage values from table data.
}
\examples{
\dontrun{
dat <- tibble(group = rep(1:5, each = 2),
              letter = c("a", "b",
                         "a", "c",
                         "a", "c",
                         "b", "e",
                         "b", "f"))

pcount(dat, letter)
pcount(dat, letter, sort = TRUE)
pcount(dat, letter, name = "obs")
}

}
