% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pfrankCopula}
\alias{pfrankCopula}
\title{Frank copulas prob}
\usage{
pfrankCopula(u1, u2, alpha, log = FALSE)
}
\arguments{
\item{u1}{dimension 1}

\item{u2}{dimension 2}

\item{alpha}{copula parameter}
}
\description{
Frank copulas prob
}
