% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pgumbelCopula}
\alias{pgumbelCopula}
\title{gumbel copulas prob}
\usage{
pgumbelCopula(u1, u2, a, give_log = FALSE)
}
\arguments{
\item{u1}{dimension 1}

\item{u2}{dimension 2}

\item{alpha}{copula parameter}
}
\description{
gumbel copulas prob
}
