% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{pick}
\alias{pick}
\title{Pick element in vector base on regex}
\usage{
pick(x, pat, ...)
}
\arguments{
\item{x}{vector}

\item{pat}{regex pattern}

\item{...}{extra args to \link[base]{grep}}
}
\description{
Pick element in vector base on regex
}
