% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{PlotData}
\alias{PlotData}
\title{PlotData function}
\usage{
PlotData(data, origin = TRUE, ...)
}
\arguments{
\item{bg}{(string) Background color Defaults to "gray90".}

\item{cols}{(string) Gridlines color  Defaults to "white".}
}
\description{
Add background and grid lines similar to ggplot.
}
\keyword{grid}
