% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottc.R
\name{plottc}
\alias{plottc}
\title{plot in terminal with color}
\usage{
plottc(
  x,
  y = NULL,
  pch = "*",
  col = "green",
  width = round(options()$width * 0.8),
  height = round(0.25 * width),
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\description{
plot in terminal with color
}
