% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_pipeline.R
\name{preprocess.genTidyDF}
\alias{preprocess.genTidyDF}
\title{Tidy a data.frame()}
\usage{
preprocess.genTidyDF(df, text_col)
}
\arguments{
\item{df}{A dataframe that will be tidied.}

\item{text_col}{The column containing the text to be tidied.}
}
\value{
A tidy dataframe.
}
\description{
This function is a wrapper for \code{unnest_tokens()} that filters emptpy tokens.
}
\examples{
\dontrun{
preprocess.genTidyDF(df, text_col)
}

}
