% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_pipeline.R
\name{preprocess.removeNonWordChars}
\alias{preprocess.removeNonWordChars}
\title{Remove Non-Word Characters}
\usage{
preprocess.removeNonWordChars(
  text,
  rm.hashtags = FALSE,
  rm.mentions = FALSE,
  rm.emoji = FALSE,
  rm.digitwords = FALSE,
  join.hyphenation = FALSE
)
}
\arguments{
\item{text}{A character vector that will be preprocessed.}

\item{rm.hashtags}{A logical, defining if #hashtags should be removed.}

\item{rm.mentions}{A logical, defining if @mentions should be removed.}

\item{rm.emoji}{A logical, defining if emoji should be removed.}

\item{rm.digitwords}{A logical, defining if all digits should be removed, including digitwords (e.g. 5G, T3, etc.)}

\item{join.hyphenation}{A logical, defining if hyphenated words should be joined.}
}
\value{
A preprocessed character vector.
}
\description{
This function preprocesses a character vector by removing non-word characters
and reports the mean number of characters before and after preprocessing.
}
\details{
By default URLs, html-entities (&nbsp), digits-words, apostrophized words, and all punctuation are removed.

Other preprocessing steps can be controlled via the arguments of the function.
}
\examples{
\dontrun{
preprocess.removeNonWordChars(
text,
rm.hashtags=FALSE,
rm.mentions=FALSE,
rm.emoji=FALSE,
rm.digitwords=FALSE,
join.hyphenation=FALSE)
}

}
