% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_pipeline.R
\name{preprocess.removeStopwords}
\alias{preprocess.removeStopwords}
\title{Remove Words From Text}
\usage{
preprocess.removeStopwords(text, stopwords)
}
\arguments{
\item{text}{A character vector that will be preprocessed.}

\item{stopwords}{A character vector containing words and phrases to be removed.}
}
\value{
A preprocessed character vector.
}
\description{
This function preprocesses a character vector by removing the words defined in \code{stopwords}.
and reports the mean number of characters before and after preprocessing.
}
\examples{
\dontrun{
stopwords <- c("in", "and", "per se")
preprocess.removeStopwords(text, stopwords)
}

}
