% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_pipeline.R
\name{preprocess.stem}
\alias{preprocess.stem}
\title{Stem word vectors}
\usage{
preprocess.stem(words, dict)
}
\arguments{
\item{words}{A character vector to be stemmed.}

\item{dict}{The hunspell dictonary to be used.}
}
\value{
A character vector of stemmed words.
}
\description{
This function is a wrapper for \code{huhunspell::huhunspell_stem()} that filters emptpy tokens.
}
\examples{
\dontrun{
preprocess.stem(words, "de_DE")
}

}
