% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pskewlogis}
\alias{pskewlogis}
\title{same as F_gllogisI}
\usage{
pskewlogis(q, scale, shape, skew)
}
\arguments{
\item{scale}{scale}

\item{shape}{shape}

\item{skew}{skewness parameters}
}
\description{
same as F_gllogisI
}
