% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptable.R
\name{ptable}
\alias{ptable}
\title{Faster \code{prop.table(table())}}
\usage{
ptable(object, percent = TRUE, sort = TRUE, ...)
}
\arguments{
\item{object}{An object to tabulate.}

\item{percent}{logical. Return rounded percentage.}

\item{sort}{logical. Return sorted table.}

\item{...}{Other arguments passed to \link[base]{table}.}
}
\value{
\code{ptable()} returns a contingency table, an object of class "table", an array of integer values. See \link[base]{table}.
}
\description{
Wrapper for \code{table()} and \code{prop.table()}. Returns APA compliant percentage values from table data.
}
\examples{
ptable(c(1,1,1,2,2,3), percent=FALSE, sort=TRUE, useNA='always')

}
