% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{py2r}
\alias{py2r}
\title{Convert ipynb to Rmarkdown or R script}
\usage{
py2r(file, name = "name", ext = "R", ...)
}
\arguments{
\item{file}{path of input}

\item{to}{R or Rmd}
}
\description{
Just a wrapper for rmarkdown and knitr
}
