% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qskewlogis}
\alias{qskewlogis}
\title{quantile function}
\usage{
qskewlogis(q, scale, shape, skew)
}
\arguments{
\item{scale}{scale}

\item{shape}{shape}

\item{skew}{skewness parameters}
}
\description{
quantile function
}
