% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{quantile95}
\alias{quantile95}
\title{quantile 95\% with name}
\usage{
quantile95(x, ...)
}
\arguments{
\item{x}{x}
}
\value{
.025. .5, and .975 quantile with name lo, up, and med
}
\description{
quantile 95\% with name
}
