% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{quartz_off}
\alias{quartz_off}
\title{Print ggplot off-screen}
\usage{
quartz_off(
  gg,
  name = "Rplot",
  width = 7,
  height = 5,
  type = "pdf",
  open = FALSE,
  ...
)
}
\arguments{
\item{gg}{ggplot object}

\item{name}{file name including path without extenstion}

\item{type}{pdf, png,...}
}
\value{
nothing
}
\description{
Print ggplot off-screen
}
