% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{query_name}
\alias{query_name}
\title{Query variable and labels from data read with haven}
\usage{
query_name(d, p, ...)
}
\arguments{
\item{d}{data}

\item{p}{character regex pattern for \link[base]{grep}}

\item{...}{extra arguments to \link[base]{grep}}
}
\description{
Query variable and labels from data read with haven
}
