% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{query_readstata13}
\alias{query_readstata13}
\title{Query data frame read with readstata13}
\usage{
query_readstata13(d, x, in_labels = TRUE)
}
\arguments{
\item{d}{Data}

\item{x}{quoted regex such as '\\d+' to find digits}

\item{in_labels}{search in labels or in var names}
}
\description{
Query data frame read with readstata13
}
