% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{r_glogisI}
\alias{r_glogisI}
\title{Sampling from skew logistic}
\usage{
r_glogisI(n, scale = 0.05, shape = 7, skew = 1)
}
\arguments{
\item{scale}{scale}

\item{shape}{shape}

\item{skew}{skewness parameters}
}
\description{
Sampling from skew logistic
}
