% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{range2seq}
\alias{range2seq}
\title{Range to sequence}
\usage{
range2seq(x)
}
\arguments{
\item{x}{vector (decimal not catched)}
}
\description{
give me a vector, I produces a continous sequence from the range
}
