% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{read_in_zip}
\alias{read_in_zip}
\title{Read a file match a pattern inside a zip file}
\usage{
read_in_zip(file, pattern, readfn = haven::read_dta)
}
\arguments{
\item{file}{the path to the zip file}

\item{pattern}{a regex pattern}

\item{readfn}{what function use to read the file?}
}
\description{
may be to improve caching file, in case one wants to read again without the 
need to extracting again
}
