% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{readnbind}
\alias{readnbind}
\title{Read and bind}
\usage{
readnbind(..., src = TRUE, agrs = NULL, fn = "readRDS", bindfn = "rbind")
}
\arguments{
\item{...}{list of paths}

\item{agrs}{extra args to fn}

\item{fn}{read function}

\item{bindfn}{bind function}

\item{xcol}{new column to identify data sources}
}
\description{
Read and bind
}
