% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilPasteboard.R
\name{readPasteboard}
\alias{readPasteboard}
\title{Read From The Pasteboard In macOS}
\usage{
readPasteboard()
}
\value{
The string currently in the pasteboard.
}
\description{
This function uses the pbpaste utility to read from the pasteboard.
}
\examples{
\dontrun{
s <- readPasteboard()
}
}
