% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigzRDS.R
\name{readRDS.gz}
\alias{readRDS.gz}
\title{Read RDS w/ pigz support}
\usage{
readRDS.gz(file, threads = (parallel::detectCores()) - 1)
}
\arguments{
\item{file}{A location to read from.}

\item{threads}{Number of threads to use. Defaults to \code{parallel::detectCores()-1}.}
}
\value{
The object at \code{file}.
}
\description{
This function uses pigz to read RDS files.
pigz is a fully functional replacement for gzip that exploits
multiple processors and multiple cores to the hilt when compressing data.
}
\details{
pigz needs to be installed for this function to work.
}
\examples{
\dontrun{
temp <- readRDS.gz('temp.RDS')
}
}
