% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{rebase}
\alias{rebase}
\title{Rebase for indexing/modellin}
\usage{
rebase(x, cpp = FALSE)
}
\arguments{
\item{x}{a vector to rebase to start from 0 or 1}

\item{cpp}{use C++ index (starting from zero instead of one)}
}
\value{

}
\description{
Rebase for indexing/modellin
}
