% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{recode_if}
\alias{recode_if}
\title{recode keeping original data when conditions have NA}
\usage{
recode_if(cond, yes, no)
}
\arguments{
\item{cond}{condition}

\item{yes}{value when condition is true}

\item{no}{value when condition is false **or NA**}
}
\description{
\code{\link[dplyr]{if_else}} will replace original data with NA when
conditions has NAs. Handling only a simple case so it is 20 times faster than
\code{\link[dplyr]{case_when}} (see Examples).
}
\examples{
x <- c(rnorm(100), NA_real_)
y <- c(rnorm(100), NA_real_)
bench::mark(
    dplyr::case_when(x < 0 ~ 10, T ~ x),
    recode_if(x < 0, 10, x),
    relative = T
)
bench::mark(
    dplyr::case_when(x < 0 ~ y, T ~ x),
    recode_if(x < 0, y, x), 
    relative = T
)
}
