% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{remove_consecutive}
\alias{remove_consecutive}
\title{Remove consecutive (by one unit only)}
\usage{
remove_consecutive(x, keep_first = TRUE)
}
\arguments{
\item{x}{vector of integer}
}
\description{
When you want to remove/merge consecutive grouping to one, such as combining consecutive survey year to one. Is there a simpler way to do this?
}
