% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_utils.R
\name{remove_text}
\alias{remove_text}
\title{Remove text from strings with gsub}
\usage{
remove_text(x, p, ...)
}
\arguments{
\item{x}{strings}

\item{p}{pattern}
}
\description{
Remove text from strings with gsub
}
