% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{rename}
\alias{rename}
\title{Rename a column}
\usage{
rename(data, old_name, new_name, ...)
}
\arguments{
\item{data}{data with columns to rename}

\item{old_name}{old name - no need to "quote"}

\item{new_name}{new name - no need to "quote"}
}
\description{
Side_0o_Effect's \url{https://stackoverflow.com/a/16490387}
}
\examples{
dt <- data.frame(a = 1, b = 2)
rename(dt, a, b)
rename(dt, a, a, unique = TRUE)
}
