% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{resize}
\alias{resize}
\title{dev.new with named width and height}
\usage{
resize(width = 7, height = 5, ...)
}
\arguments{
\item{width}{width}

\item{height}{height}
}
\description{
dev.new with named width and height
}
