% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{rmvnorm_sparseprec}
\alias{rmvnorm_sparseprec}
\title{Sample multivariates with precision matrix}
\usage{
rmvnorm_sparseprec(n, mean = rep(0, nrow(prec)), prec = diag(length(mean)))
}
\arguments{
\item{n}{number of samples}

\item{mean}{scalar/vector of the mean(s)}

\item{prec}{precision matrix}
}
\value{
n samples
}
\description{
Sample multivariates with precision matrix
}
