% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{run_tmb_on_slurm}
\alias{run_tmb_on_slurm}
\title{Run TMB Model on SLURM Cluster}
\usage{
run_tmb_on_slurm(
  data,
  params,
  random = NULL,
  map_words = NULL,
  dll = "model",
  n_cores = NULL,
  job_name = "TMBjob",
  partition = "fuchs",
  time = "8:00:00",
  ntasks = 1,
  nodes = 1,
  conda_env = "kinh",
  rscript = "run_tmb.r",
  work_dir = ".",
  submit = FALSE,
  submit_cmd = "sbatch",
  data_path = "tmb_data.RData",
  param_path = "tmb_params.RData",
  map_path = "tmb_param_map.RData"
)
}
\arguments{
\item{data}{Data list for TMB model}

\item{params}{Parameter list for TMB model}

\item{random}{Character vector of random effect parameter names}

\item{map_words}{Named list of parameter mapping specifications}

\item{dll}{Character name of compiled TMB model (without extension)}

\item{n_cores}{Number of cores for OpenMP parallelization}

\item{job_name}{Character name for SLURM job}

\item{partition}{Character name of SLURM partition}

\item{time}{Character time limit for job ("HH:MM:SS")}

\item{ntasks}{Number of tasks for SLURM}

\item{nodes}{Number of nodes for SLURM}

\item{conda_env}{Character name of conda environment}

\item{rscript}{Character name for generated R script}

\item{work_dir}{Character path to working directory}

\item{submit}{Logical whether to submit job immediately}

\item{submit_cmd}{Character command to submit SLURM job}

\item{data_path}{Character path to save data}

\item{param_path}{Character path to save parameters}

\item{map_path}{Character path to save parameter map}
}
\value{
Invisibly returns a list with:
  \item{submitted}{Logical indicating if job was submitted}
  \item{output}{System output from job submission if applicable}
  \item{rscript}{Path to generated R script}
  \item{slurm}{Path to generated SLURM script}
  \item{data}{Path to saved data}
  \item{param}{Path to saved parameters}
  \item{map}{Path to saved parameter map if applicable}
}
\description{
Creates and optionally submits a SLURM job to run a TMB model on a computing cluster.
}
\details{
This function generates necessary R and SLURM scripts to run a TMB model
on a SLURM-based computing cluster. It saves model data and parameters
to files, creates an R script to fit the model, and a SLURM submission
script. The job can be submitted immediately or the scripts can be
saved for later submission.
}
