% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R, R/plotUtils.R
\name{s2n}
\alias{s2n}
\title{s2n - String to numeric}
\usage{
s2n(x = "string")

s2n(x = "string")
}
\description{
Converting string to ACSII number, to use with plotting parameter `pch`

Converting string to ACSII number, to use with plotting parameter `pch`
}
