% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{save_jpg_for_submission}
\alias{save_jpg_for_submission}
\title{ggplot to 300 dpi JPG for article submission}
\usage{
save_jpg_for_submission(name, figure = NULL, ...)
}
\arguments{
\item{name}{name to save without extension}

\item{figure}{ggplot object, empty means the last plot}

\item{...}{more parameters for `ggsave`}
}
\description{
what this does is save as PDF, then crop it, and convert to JPG which
}
\details{
- not change scale, as compared to `ggsave` as JPG and put `dpi = 300`, always set width and height! my favorites are 7-4, 7-5, 9-6
- crop white space surrounding (such as from `geom_sf`), with `pdfcrop` or `pdfcrop.pl` for mamba/conda.
- convert to JPG with `pdftoppm` (from `poppler-utils`)

this will create a subfolder at `here::here('fig')` and save into it.

Try adding `+ k.theme`?
}
