% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{savePDF}
\alias{savePDF}
\title{savePDF function}
\usage{
savePDF(name = "name", w = dev.size()[1], h = dev.size()[2])
}
\arguments{
\item{bg}{(string) Background color Defaults to "gray90".}

\item{cols}{(string) Gridlines color  Defaults to "white".}
}
\description{
Add background and grid lines similar to ggplot.
}
\examples{
savePDF()
}
\keyword{grid}
