% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{savePs}
\alias{savePs}
\title{savePs function}
\usage{
savePs(name = "name", w = dev.size()[1], h = dev.size()[2])
}
\arguments{
\item{bg}{(string) Background color Defaults to "gray90".}

\item{cols}{(string) Gridlines color  Defaults to "white".}
}
\description{
Add background and grid lines similar to ggplot.
}
\keyword{grid}
