% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{screen_to_file}
\alias{screen_to_file}
\title{Open output of a function in a new text file (default application)}
\usage{
screen_to_file(x)
}
\arguments{
\item{x}{name of a function}
}
\description{
Open output of a function in a new text file (default application)
}
\examples{
fn_to_file(lm)
}
