% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_timestamp.R
\name{seq_timestamp}
\alias{seq_timestamp}
\title{Sequence of Timestamps}
\usage{
seq_timestamp(n = 2)
}
\arguments{
\item{n}{the desired output vector length.}
}
\value{
a numeric vector of timestamps.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\dontrun{
seq_timestamp(n = 2)}
}
