% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ui.R
\name{simple_ui}
\alias{simple_ui}
\title{Shiny Outputs}
\usage{
simple_ui(id, output, type = "ui", ...)
}
\arguments{
\item{id}{the id of the module.}

\item{output}{the name of the output.}

\item{type}{the type of output: "ui" (default), "text", "plot", "table".}

\item{...}{other arguments to pass to the *Output function.}
}
\value{
an HTML tag.
}
\description{
Wrapper around Shiny *Output functions.
}
\details{
\code{id} can be a vector of ids in case of sub module.
}
\examples{
simple_ui(id = "my_module", output = "my_output")
}
