% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{sizing}
\alias{sizing}
\title{Sizing the figure in Python notebook IRKernel}
\usage{
sizing(w = 7, h = 4)
}
\arguments{
\item{w}{repr.plot.width}

\item{h}{repr.plot.height}
}
\description{
too lazy to type the long options?
}
\examples{
\dontrun{
sizing(7, 4)
}
}
