% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{skim}
\alias{skim}
\title{Simpler data overview}
\usage{
skim(x, ncol = 5)
}
\arguments{
\item{x}{data frame}

\item{ncol}{wrap the variables description into several columns to print}
}
\description{
Simpler data overview
}
