\name{smidR-package}
\alias{smidR-package}
\alias{smidR}
\docType{package}
\title{
\packageTitle{smidR}
}
\description{
\packageDescription{smidR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{smidR}
\packageIndices{smidR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{smidR}

Maintainer: \packageMaintainer{smidR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
