% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_utils.R
\name{snaking}
\alias{snaking}
\title{Convert to snake case from package snakecase}
\usage{
snaking(
  string,
  abbreviations = NULL,
  sep_in = "[^[:alnum:]]",
  parsing_option = 1,
  transliterations = NULL,
  numerals = "middle",
  sep_out = NULL,
  unique_sep = NULL,
  empty_fill = NULL,
  prefix = "",
  postfix = ""
)
}
\description{
Convert to snake case from package snakecase
}
