% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_all.R
\name{source_all}
\alias{source_all}
\title{Source All Code}
\usage{
source_all(path, verbose = TRUE)
}
\arguments{
\item{path}{a character vector of the path to explore}

\item{verbose}{a logical value (default = TRUE). If FALSE, then minimal trace will be sent to the console}
}
\value{
an integer of how many functions have been attached to the environment
}
\description{
Source All Code
}
\details{
Given path is explored using list.files with "*.R" pattern, and recursive = TRUE
All listed files are loaded using source()
}
\examples{
\dontrun{
source_all("./R")
}
}
