% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.r
\name{strip_U}
\alias{strip_U}
\title{Strip unicode characters}
\usage{
strip_U(x)
}
\arguments{
\item{x}{string with unicode such as "<U+2019>"}
}
\description{
the stripped one can be used in, for example, selecting columns.
}
