% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.igraph.R
\name{summary.igraph}
\alias{summary.igraph}
\title{A better summary function for igraph graphs}
\usage{
\method{summary}{igraph}(object, ...)
}
\arguments{
\item{object}{An object to summarise.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
A summary-function for igraph objects.
}
\examples{
\dontrun{
summary(g)
}
}
