% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{surv_split}
\alias{surv_split}
\title{Split time to interval for survival model - improved}
\usage{
surv_split(
  x,
  duration,
  event,
  cuts,
  time_varying = NULL,
  label_episode = FALSE
)
}
\arguments{
\item{x}{data frame}

\item{duration}{exposure time}

\item{event}{binary 0 and 1 coded}

\item{time_varying}{not in use}

\item{label_episode}{set to TRUE to have the episode labeled with actual cut
intervals, otherwise episodes are integers}

\item{cut}{the cut points on duration forming intervals [min_duration, first_value],
(first_value, second_value], ..., (last_value, max_duration]}
}
\description{
This is similar to \code{\link[survival]{survSplit}} but allow the starting
time and ending time to be equal (e.g., immediate death, automatic default in
loan) and automatically include min and max of duration in the cut intervals
}
\details{
The tested speed is 1.37 times to that of survSplit.
}
\examples{
x <- data.frame(time = c(5, 10), event = c(1, 0))
surv_split(x, "time", "event", c(2,4,6))
surv_split(x, "time", "event", c(2,4,6), label_episode = TRUE)
}
