% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{table}
\alias{table}
\title{tabulate a variable with pipe}
\usage{
table(.data, var, useNA = "a", ...)
}
\arguments{
\item{.data}{data containing the variable to tabulate}

\item{var}{the variable to tabulate}

\item{...}{
  Arguments passed on to \code{\link[base:table]{base::table}}
  \describe{
    \item{\code{exclude}}{levels to remove for all factors in \code{\dots}.  If
    it does not contain \code{\link[base]{NA}} and \code{useNA} is not
    specified, it implies \code{useNA = "ifany"}.  See
    \sQuote{Details} for its interpretation for non-factor arguments.}
    \item{\code{useNA}}{whether to include \code{NA} values in the table.
    See \sQuote{Details}.  Can be abbreviated.}
    \item{\code{dnn}}{the names to be given to the dimensions in the result (the
    \emph{dimnames names}).}
    \item{\code{deparse.level}}{controls how the default \code{dnn} is
    constructed.  See \sQuote{Details}.}
  }}
}
\description{
default table is easier to read than \code{\link[dplyr]{count}}
}
