% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{table_as_html}
\alias{table_as_html}
\title{View table as html}
\usage{
table_as_html(code, ...)
}
\arguments{
\item{code}{out of knitr::kable(format='html') for}
}
\description{
for, e.g, copy to Word editors keeping most of the format
}
