% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{take_note}
\alias{take_note}
\title{Write a note to file}
\usage{
take_note(
  text = "text",
  to = "Rnote",
  thisfilename = NULL,
  home = FALSE,
  wd = !home,
  time_stamp = TRUE
)
}
\description{
Write a note to file
}
