% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{tape}
\alias{tape}
\title{tape list data.frame}
\usage{
tape(fn, a_list, ...)
}
\arguments{
\item{fn}{function to apply to each list}

\item{a_list}{a list of data frame}

\item{...}{extra args to Map}
}
\description{
tape list data.frame
}
\examples{
parent <- list(
   list(
       data.frame(lds=1:10, q = 10:1), 
       data.frame(lds=1:10, q = 10:1)
   ),
   list(
       data.frame(lds=1:10, q = 10:1), 
       data.frame(lds=1:10, q = 10:1)
   )
)
tape(function(x, y) cbind(x, y), l[[1]], y='extra_col')
tape(function(child, replicate) 
   tape(function(grandchild, model) cbind(grandchild, replicate, model), 
       child, model = seq_along(child)), 
       parent, replicate = seq_along(parent))
}
