% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{tbl_AB}
\alias{tbl_AB}
\title{Tabulate 2 variables with count and percentage}
\usage{
tbl_AB(.data, row, col, digits = 2)
}
\arguments{
\item{.data}{data}

\item{row}{row variable}

\item{col}{column variable}

\item{digits}{round the percentages}
}
\description{
Tabulate 2 variables with count and percentage
}
