% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_kudusch.R
\name{theme_kudusch}
\alias{theme_kudusch}
\title{A simple custom theme for ggplot.}
\usage{
theme_kudusch(base_size = 9, base_family = "Anka/Coder Narrow")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A minimal theme based on \code{theme_classic} with a monospaced font.
}
\examples{
\dontrun{
mtcars2 <- within(mtcars, {
vs <- factor(vs, labels = c("V-shaped", "Straight"))
am <- factor(am, labels = c("Automatic", "Manual"))
cyl  <- factor(cyl)
gear <- factor(gear)
})
p1 <- ggplot(mtcars2) +
geom_point(aes(x = wt, y = mpg, colour = gear)) +
labs(title = "Fuel economy declines as weight increases",
subtitle = "(1973-74)",
caption = "Data from the 1974 Motor Trend US magazine.",
tag = "Figure 1",
x = "Weight (1000 lbs)",
y = "Fuel economy (mpg)",
colour = "Gears")

p1 + theme_kudusch()
# Theme examples with panels
p2 <- p1 + facet_grid(vs ~ am)
p2 + theme_kudusch()
}

}
