% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{tmb_compile_and_load}
\alias{tmb_compile_and_load}
\title{TMB compile and load}
\usage{
tmb_compile_and_load(code)
}
\arguments{
\item{x}{model character}
}
\value{
dll
}
\description{
TMB compile and load
}
