% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{tmb_fixit}
\alias{tmb_fixit}
\title{TMB fix parameters}
\usage{
tmb_fixit(par, fix_names)
}
\arguments{
\item{par}{initial set of parameter}

\item{fix_names}{names of parameters to be fixed (not fitting)}
}
\value{
TMB's format for map argument
}
\description{
TMB fix parameters
}
\seealso{
\code{\link[ktools]{get_report}}.
}
