% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{tmb_load}
\alias{tmb_load}
\title{Basic loading of TMB model with basic name handling, unloading}
\usage{
tmb_load(x, ...)
}
\arguments{
\item{x}{character name of the TMB model, with or without extension is OK}

\item{...}{Variables to set}
}
\description{
Basic loading of TMB model with basic name handling, unloading
}
