% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{tmb_unload}
\alias{tmb_unload}
\title{Unload TMB if loaded}
\usage{
tmb_unload(x)
}
\arguments{
\item{x}{character name of the TMB file}
}
\value{
nothing
}
\description{
help when rerun the code with unload included
}
