% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toupper_words.R
\name{toupper_words}
\alias{toupper_words}
\title{Capitalize First Letters}
\usage{
toupper_words(s, strict = FALSE)
}
\arguments{
\item{s}{a character string.}

\item{strict}{a logical (default = FALSE) if other characters should be forced to lower.}
}
\value{
a character string.
}
\description{
Capitalize First Letters
}
\examples{
toupper_words("this is an example")
}
