% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toupperfirst.R
\name{toupperfirst}
\alias{toupperfirst}
\title{Capitalize First Letter}
\usage{
toupperfirst(x)
}
\arguments{
\item{x}{a character vector, or an object that can be coerced to character by as.character.}
}
\value{
a character vector.
}
\description{
Capitalize First Letter
}
\examples{
toupperfirst("capitalize first letter")
toupperfirst(1)
}
