% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_level.R
\name{trace_level}
\alias{trace_level}
\title{Trace Level}
\usage{
trace_level(level = NULL)
}
\arguments{
\item{level}{an integer (or NULL).}
}
\value{
Nothing or the active trace level.
}
\description{
A helper function to get/set the trace level printed by \code{\link[=catl]{catl()}}.
}
\details{
\code{level} is used to define the level of trace printed by \code{\link[=catl]{catl()}}.
Use 0 to set the trace OFF or NULL to get the current trace level.
}
\examples{
trace_level()
trace_level(2)
trace_level(0)
}
\seealso{
\code{\link[=catl]{catl()}}
}
