% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{unkount}
\alias{unkount}
\title{Uncounting data frame using a weights}
\usage{
unkount(x, weight)
}
\arguments{
\item{x}{a data frame}

\item{weight}{column in x with number of replicates, no need to "quote"}
}
\description{
Similar to \code{\link[tidyr]{uncount}} but using base R only
}
\examples{
x <- data.frame(a = c(1, 2), b = c(3, 4))
unkount(x, b)
}
