% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_r_environ.R
\name{update_r_environ}
\alias{update_r_environ}
\title{Update .Renviron File}
\usage{
update_r_environ(path = getwd(), key, value)
}
\arguments{
\item{path}{the path to the .Renviron file (default is working directory).}

\item{key}{a character string for the name of the variable.}

\item{value}{the value to be assigned to the variable.}
}
\value{
a logical (TRUE if success, else FALSE)
}
\description{
Add an environment variable to the project.
}
\examples{
\dontrun{
update_r_environ(key = "MY_VAR", value = "my_value")
}
}
\seealso{
\code{\link[=use_r_environ]{use_r_environ()}}
}
