% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_plot_theme.R
\name{use_plot_theme}
\alias{use_plot_theme}
\title{Use Plot Theme}
\usage{
use_plot_theme(path = getwd())
}
\arguments{
\item{path}{where to copy the template.}
}
\value{
a logical.
}
\description{
Create a theme function that can be used across a project.
}
\examples{
\dontrun{
use_plot_theme()
}
}
