% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_environ.R
\name{use_r_environ}
\alias{use_r_environ}
\title{Create .Renviron File}
\usage{
use_r_environ(path = getwd())
}
\arguments{
\item{path}{the path where to create the file.}
}
\value{
a logical (see \code{\link[=file.create]{file.create()}})
}
\description{
Create .Renviron File
}
\details{
By default, \code{path} is set to the working directory.
}
\examples{
\dontrun{
use_r_environ()
}
}
\seealso{
\code{\link[=update_r_environ]{update_r_environ()}}
}
