% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_template.R
\name{use_template}
\alias{use_template}
\title{Copy Template File}
\usage{
use_template(template, name = template, package = "ktools", path = getwd())
}
\arguments{
\item{template}{the name of the template file to look for}

\item{name}{the name of the target file (default: same as template)}

\item{package}{the name of the package (to pass to system.file, default = "ktools")}

\item{path}{the path where to copy the template file (default: working directory)}
}
\description{
Copy Template File
}
\examples{
\dontrun{
use_template(template = "global.R")
}
}
