% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_adj.R
\name{var_ll}
\alias{var_ll}
\title{Variance of adjusted prevalence using sensitivity and specificity}
\usage{
var_ll(pi, n, spec = 0.9, sens = 0.995)
}
\arguments{
\item{pi}{the adjusted prevalence}

\item{n}{sample size}

\item{spec}{specificity}

\item{sens}{sensitivity}
}
\description{
using directly the formula not posterior distribution
}
