% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{var_skew_llogis}
\alias{var_skew_llogis}
\title{Compute the variance of skew loglogistic distribution}
\usage{
var_skew_llogis(shape, skew)
}
\arguments{
\item{shape}{shape}

\item{skew}{skewness parameters}

\item{scale}{scale}
}
\description{
Compute the variance of skew loglogistic distribution
}
