% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{wait}
\alias{wait}
\title{wait function}
\usage{
wait(wait.time = 2)
}
\arguments{
\item{bg}{(string) Background color Defaults to "gray90".}

\item{cols}{(string) Gridlines color  Defaults to "white".}
}
\description{
Add background and grid lines similar to ggplot.
}
\examples{
wait()
}
\keyword{grid}
