% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{which_in}
\alias{which_in}
\title{Find element in a vector and extract name not indice}
\usage{
which_in(x, y)
}
\arguments{
\item{x}{a vector (scalar is also evaluated to TRUE)}

\item{y}{a vector}
}
\description{
Find element in a vector and extract name not indice
}
