% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilPasteboard.R
\name{writePasteboard}
\alias{writePasteboard}
\title{Write To The Pasteboard In macOS}
\usage{
writePasteboard(s)
}
\arguments{
\item{s}{An character string to write to the pasteboard.}
}
\description{
This function uses the pbcopy utility to write to the pasteboard.
}
\examples{
\dontrun{
writePasteboard("Hello!")
}

}
