% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catl.R
\name{catl}
\alias{catl}
\title{Wrapper Around Cat}
\usage{
catl(..., level = 1, debug = Sys.getenv("TRACE_LEVEL"), newline = TRUE)
}
\arguments{
\item{...}{R objects as defined in \link[base]{cat}}

\item{level}{a numeric value indicating the level of trace}

\item{debug}{an optional (usually omitted) value to indicate the level of trace to be displayed}

\item{newline}{a logical. If TRUE (default), new line \code{"\n"} will be added}
}
\value{
None (invisible NULL)
}
\description{
This wrapper function aims at enabling a trace level mechanism that relies
on the \link[base]{cat} function.
}
\details{
The goal is to provide a trace level mechanism that can be disabled in a
production context (as printing output to the console or file is costly).

The different levels are left flexible.

By default, \code{debug} is set to \code{Sys.getenv("TRACE_LEVEL")}
If the environment variable is not set, no output will be produced (silent mode)
When the variable is set, the function will pass calls with level <= debug
to the original \link[base]{cat} function.

The \link[ktools]{trace_level} function can be used to activate a specific trace level.
}
\examples{
# silent unless TRACE_LEVEL is set
catl("Awesome message here")

# force print no matter if TRACE_LEVEL is set
catl("Awesome message here", debug = 1)

# demonstrate trace levels
catl("Awesome message here", level = 2, debug = 1)
catl("Awesome message here", level = 1, debug = 2)

}
