% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_template.R
\name{copy_template}
\alias{copy_template}
\title{Copy Package Template Files}
\usage{
copy_template(template, path = getwd(), filename = NULL)
}
\arguments{
\item{template}{the name of the template file to be copied}

\item{path}{destination path where to copy the template}

\item{filename}{the name of the copy}
}
\value{
the output of the file.copy function call
}
\description{
Copy Package Template Files
}
\examples{
copy_template(template = "template_shiny_server", path = "shinyapp", filename = "my_server.R")
}
