% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_range.R
\name{date_range}
\alias{date_range}
\title{Compute date range}
\usage{
date_range(min, max, type = "this_year")
}
\arguments{
\item{min}{a Date for the lower value}

\item{max}{a Date for the higher value}

\item{type}{an optional character vector, to set the strategy (default = "this_year")}
}
\value{
a Date vector c(start, end)
}
\description{
Compute specific date range between two dates.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
date_range(min = Sys.Date()-365, max = Sys.Date()+365, type = "this_year")
}
