% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_range.R
\name{date_range}
\alias{date_range}
\title{Compute date range}
\usage{
date_range(min, max, type = "this_year")
}
\arguments{
\item{min}{a Date for the lower value}

\item{max}{a Date for the higher value}

\item{type}{an optional character vector, to set the strategy (default = "this_year")}
}
\value{
a Date vector c(min, max)
}
\description{
Compute date range
}
\examples{
date_range(min = Sys.Date()-365, max = Sys.Date()+365, type = "this_year")
}
