% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dskewlogis}
\alias{dskewlogis}
\title{same as f_gllogisI}
\usage{
dskewlogis(x, scale, shape, skew, log = FALSE)
}
\arguments{
\item{scale}{scale}

\item{shape}{shape}

\item{skew}{skewness parameters}
}
\description{
same as f_gllogisI
}
