% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{facet_unwrap}
\alias{facet_unwrap}
\title{Unwrap a facet_wrap plot}
\usage{
facet_unwrap(g, plot = TRUE, save = FALSE, path = ".", ...)
}
\arguments{
\item{g}{the ggplot2 object created as normal}

\item{plot}{whether to plot (or just save by turn the next args to TRUE)}

\item{save}{whether to save the plot to disk}

\item{path}{where to save them?}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggsave]{ggplot2::ggsave}}
  \describe{
    \item{\code{filename}}{File name to create on disk.}
    \item{\code{device}}{Device to use. Can either be a device function
(e.g. \link{png}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only). If
\code{NULL} (default), the device is guessed based on the \code{filename} extension.}
    \item{\code{scale}}{Multiplicative scaling factor.}
    \item{\code{width,height}}{Plot size in units expressed by the \code{units} argument.
If not supplied, uses the size of the current graphics device.}
    \item{\code{units}}{One of the following units in which the \code{width} and \code{height}
arguments are expressed: \code{"in"}, \code{"cm"}, \code{"mm"} or \code{"px"}.}
    \item{\code{dpi}}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}
    \item{\code{limitsize}}{When \code{TRUE} (the default), \code{ggsave()} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
    \item{\code{bg}}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}
    \item{\code{create.dir}}{Whether to create new directories if a non-existing
directory is specified in the \code{filename} or \code{path} (\code{TRUE}) or return an
error (\code{FALSE}, default). If \code{FALSE} and run in an interactive session,
a prompt will appear asking to create a new directory when necessary.}
  }}
}
\value{
invisible a list of ggplot objects, one for each panels
}
\description{

}
\examples{
\dontrun{
g <- iris \%>\%
  ggplot(aes(Sepal.Length)) +
  geom_histogram() +
  facet_wrap(~Species)
facet_unwrap(g)
}
}
