% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{facet_unwrap}
\alias{facet_unwrap}
\title{Unwrap a facet_wrap plot}
\usage{
facet_unwrap(g, plot = TRUE, save = FALSE, path = ".", ...)
}
\arguments{
\item{g}{the ggplot2 object created as normal}

\item{plot}{whether to plot (or just save by turn the next args to TRUE)}

\item{save}{whether to save the plot to disk}

\item{path}{where to save them?}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggsave]{ggplot2::ggsave}}
  \describe{
    \item{\code{filename}}{File name to create on disk.}
    \item{\code{device}}{Device to use. Can either be a device function
(e.g. \link{png}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}
    \item{\code{scale}}{Multiplicative scaling factor.}
    \item{\code{width}}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}
    \item{\code{height}}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}
    \item{\code{units}}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}
    \item{\code{dpi}}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}
    \item{\code{limitsize}}{When \code{TRUE} (the default), \code{ggsave()} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
    \item{\code{bg}}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}
  }}
}
\description{

}
\examples{
\dontrun{
g <- iris \%>\%
  ggplot(aes(Sepal.Length)) +
  geom_histogram() +
  facet_wrap(~Species)
facet_unwrap(g)
}
}
