% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_date.R
\name{filter_date}
\alias{filter_date}
\title{Filter by Date Range}
\usage{
filter_date(x, colname = NULL, ref = Sys.Date(), unit = "month")
}
\arguments{
\item{x}{a data.frame with a Date or POSIXct column}

\item{colname}{an optional character string to provide the name of the column to filter}

\item{ref}{an optional Date value to be used as a reference for the date range}

\item{unit}{a string (see \code{lubridate::round_date()})}
}
\value{
a filtered data.frame
}
\description{
Filter by Date Range
}
\examples{
# -- this month (the default)
filter_date(data.frame(date = Sys.Date() - runif(n = 10, min = 1, max = 50)))

# -- previous month
filter_date(data.frame(date = Sys.Date() - runif(n = 10, min = 1, max = 50)), ref = Sys.Date() - 31)

}
