% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rank_diff.R
\name{get_rank_diff}
\alias{get_rank_diff}
\title{Calc rank diff}
\usage{
get_rank_diff(tidy_x, tidy_y, id = "id", token = "token")
}
\arguments{
\item{tidy_x}{A data frame created by tidytext::unnest_tokens()}

\item{tidy_y}{A data frame created by tidytext::unnest_tokens()}

\item{id}{string. Name of id variable. Defaults to "id".}

\item{token}{string. Name of token variable. Defaults to "token".}
}
\value{
Returns a data frame.
}
\description{
Lorem upsim
}
\examples{
\dontrun{
rank_diff <- get_rank_diff(tidy_x, tidy_y, id = "id", token = "token")
}

}
