% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_date.R
\name{has_date}
\alias{has_date}
\title{Find Date Column}
\usage{
has_date(x, single = TRUE)
}
\arguments{
\item{x}{a data.frame object}

\item{single}{a logical (default TRUE) if it should return only the first column}
}
\value{
the name of the column(s) or NA if no date column is found
}
\description{
Find Date Column
}
\examples{
has_date(data.frame(date = Sys.Date()))
has_date(data.frame(date = Sys.Date(), update = Sys.Date()), single = FALSE)
has_date(data.frame(date = Sys.Date(), update = Sys.Date()), single = TRUE)
}
