% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incat.R
\name{incat}
\alias{incat}
\title{Indented Cat}
\usage{
incat(x, sep = "")
}
\arguments{
\item{x}{R objects (see cat function for details)}

\item{sep}{a character vector of strings to append after each element (default is "").}
}
\value{
None (invisible NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
When the function detects a direct call (i.e. not from another function), then it basically just do cat
otherwise it finds the calling function name and print indented output based on callstack position.
}
\examples{
\dontrun{
incat("This message")
}
}
