% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_decode.R
\name{input_decode}
\alias{input_decode}
\title{Decode Input Value}
\usage{
input_decode(x)
}
\arguments{
\item{x}{the (input) value to decode.}
}
\value{
a named vector c(namespace, action, value) or c(action, value)
}
\description{
Split an input value with pattern "namespace-action_value" into a named vector.
}
\details{
The function is used together with the \code{\link[=action_link]{action_link()}} / \code{\link[=onclick_event]{onclick_event()}} functions.
}
\examples{
input_decode("module1-add_leg_896")
input_decode("add_leg_896")
}
