% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktools-package.R
\docType{package}
\name{ktools-package}
\alias{ktools}
\alias{ktools-package}
\title{ktools: Utility Tools}
\description{
Various utility tools to make life easier and homogenize patterns across projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thekangaroofactory.github.io/ktools/}
  \item Report bugs at \url{https://github.com/thekangaroofactory/ktools/issues}
}

}
\author{
\strong{Maintainer}: Philippe Peret \email{philippe.peret@hotmail.com} (\href{https://orcid.org/0009-0003-9666-7490}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Philippe Peret \email{philippe.peret@hotmail.com} (\href{https://orcid.org/0009-0003-9666-7490}{ORCID}) [copyright holder]
}

}
\keyword{internal}
