% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{MakeADFunSafe}
\alias{MakeADFunSafe}
\title{MakeADFun safely terminated if there is bound error}
\usage{
MakeADFunSafe(...)
}
\arguments{
\item{...}{MakeADFun argments}
}
\description{
to prevent TMB crashing an R session
}
