% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.list.R
\name{match.list}
\alias{match.list}
\title{Match Lists}
\usage{
match.list(x, reference)
}
\arguments{
\item{x}{a named list to check.}

\item{reference}{a named list to use as a reference.}
}
\value{
a matched list
}
\description{
Matches a named list against a reference.
}
\details{
The function will drop any element from \code{x} that does not fit with a name
found in \code{reference} and add missing ones.

It will also check that elements in \code{x} have same classes as matching element in \code{reference}.
}
\examples{
match.list(x = list(foo = 1), reference = list(foo = 0, bar = "dummy"))
}
