% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onclick_event.R
\name{onclick_event}
\alias{onclick_event}
\title{Build onClick Event}
\usage{
onclick_event(target, value = NULL, namespace = NULL)
}
\arguments{
\item{target}{the input to update on click}

\item{value}{the value to send to the input}

\item{namespace}{the id of the module to target if so}
}
\value{
a string to pass to an onclick parameter
}
\description{
This function helps building onclick instruction for a listener
to apply / decode.
This is particularly usefull when working with lists of links, most probably
generated within an apply instruction over a data.frame.
}
\details{
\code{target} defines the name of the input entry to update (i.e. input$target).

When \code{value} is NULL (default), the id of the html tag on which the onClick
is applied will be sent.
This enables both strategies:
\itemize{
\item send the id of the tag that has been clicked
\item send an instruction for the target to decode (see examples)
}
}
\examples{
# -- basic use case:
# input$btn_clicked will get the id of the html tag
onclick_event(target = "btn_clicked")

# -- custom value:
# input$btn_clicked will get "foo" each time the tag is clicked
onclick_event(target = "btn_clicked", value = "foo")

# -- shiny module:
# input$btn_clicked inside "module" will get 12 each time the tag is clicked
onclick_event(target = "btn_clicked", value = 12, namespace = "module")
}
