% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_table.R
\name{p_table}
\alias{p_table}
\title{Faster prop.table(table())}
\usage{
p_table(object, ...)
}
\arguments{
\item{object}{An object to tabulate.}

\item{...}{Other arguments passed to \link[base]{table}.}
}
\description{
Faster prop.table(table())
}
\examples{
p_table(df)
}
