% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigzRDS.R
\name{readRDS.gz}
\alias{readRDS.gz}
\title{Read RDS w/ pigz support}
\usage{
readRDS.gz(file, threads = (parallel::detectCores()) - 1)
}
\arguments{
\item{file}{A location to read from.}
}
\value{
The object at \code{file}.
}
\description{
Read RDS w/ pigz support
}
\examples{
temp <- readRDS.gz('temp.RDS')
}
