% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigzRDS.R
\name{saveRDS.gz}
\alias{saveRDS.gz}
\title{Save RDS w/ pigz support}
\usage{
saveRDS.gz(object, file, threads = (parallel::detectCores()) - 1,
  compression_level = 6)
}
\arguments{
\item{object}{An object to save.}

\item{file}{A location to save to.}
}
\description{
Save RDS w/ pigz support
}
