% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_code.R
\name{scan_code}
\alias{scan_code}
\title{Scan Code}
\usage{
scan_code(path = getwd())
}
\arguments{
\item{path}{a path to scan}
}
\value{
a data.frame with counts for the scanned files
}
\description{
Generates a code report for a given path.
}
\examples{
scan_code()
}
\seealso{
\code{\link[=count_lines]{count_lines()}}
}
