% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{screen_to_file}
\alias{screen_to_file}
\title{Open output of a function in a new text file (in default application)}
\usage{
screen_to_file(x)
}
\arguments{
\item{x}{name of a function}
}
\description{
Useful to run through a function source code for understanding, debugging
}
\examples{
\dontrun{
screen_to_file(lm)
}
}
