% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_screenshot.R
\name{screenshot_start}
\alias{screenshot_start}
\title{Start Record Plots}
\usage{
screenshot_start(height = NA, width = NA, dpi = 300)
}
\arguments{
\item{height}{plot size (units)}

\item{width}{plot size (units)}

\item{dpi}{plot resolution value (default = 300)}
}
\value{
NULL (invisible)
}
\description{
This is a wrapper function to init plot viewer recording.
}
\details{
The camcorder package is set as suggest in this package to avoid unnecessary
installation or extra dependencies. In case it is not found, the function
will throw a warning and exit with NULL.
}
\examples{
\dontrun{
screenshot_start(height = 5, width = 10, dpi = 350)
}
}
\seealso{
\code{\link[camcorder:Recording]{camcorder::gg_record()}}
}
