% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_screenshot.R
\name{screenshot_start}
\alias{screenshot_start}
\title{Start Record Plots}
\usage{
screenshot_start(height = NULL, width = NULL, dpi = 300)
}
\arguments{
\item{height}{plot size in units (see camcorder::gg_record())}

\item{width}{plot size in units (see camcorder::gg_record())}

\item{dpi}{plot resolution (see camcorder::gg_record())}
}
\description{
Start Record Plots
}
\examples{
\dontrun{
screenshot_start(height = 5, width = 10, dpi = 350)
}
}
