% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_screenshot.R
\name{screenshot_stop}
\alias{screenshot_stop}
\title{Stop Record Plots}
\usage{
screenshot_stop(
  first = 1,
  last = 1,
  frame = 0.25,
  loop = TRUE,
  background = "none"
)
}
\arguments{
\item{first}{see camcorder::gg_playback()}

\item{last}{see camcorder::gg_playback()}

\item{frame}{see camcorder::gg_playback()}

\item{loop}{logical (default = TRUE)}

\item{background}{character string for the background color}
}
\value{
nothing. Used for side effect.
}
\description{
This is a wrapper function to stop plot viewer recording and build a gif
from the captures.
}
\details{
The camcorder package is set as suggest in this package to avoid unnecessary
installation or extra dependencies. In case it is not found, the function
will throw a warning and exit with NULL.
}
\examples{
\dontrun{
screenshot_stop(first = 1, last = 1, frame = 0.25, loop = TRUE, background = "none")
}
}
\seealso{
\code{\link[camcorder:Recording]{camcorder::gg_playback()}}
}
