% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_screenshot.R
\name{screenshot_stop}
\alias{screenshot_stop}
\title{Stop Record Plots}
\usage{
screenshot_stop(
  first = 1,
  last = 1,
  frame = 0.25,
  loop = TRUE,
  background = "none"
)
}
\arguments{
\item{first}{see camcorder::gg_playback()}

\item{last}{see camcorder::gg_playback()}

\item{frame}{see camcorder::gg_playback()}

\item{loop}{logical (default = TRUE)}

\item{background}{character string for the background color}
}
\description{
Stop Record Plots
}
\examples{
\dontrun{
screenshot_stop(first = 1, last = 1, frame = 0.25, loop = TRUE, background = "none")
}
}
