% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_timestamp.R
\name{seq_timestamp}
\alias{seq_timestamp}
\title{Generate a sequence of timestamps}
\usage{
seq_timestamp(n = 2)
}
\arguments{
\item{n}{an integer representing the desired output vector length (see details)}
}
\value{
a numeric vector of timestamps
}
\description{
Generate a sequence of timestamps
}
\details{
n is expected to receive an integer. If a decimal value is provided, then round()
will be used, which means that 2.8 will return a length 3 vector.
}
\examples{
seq_timestamp(n = 2)
}
