% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ui.R
\name{simple_ui}
\alias{simple_ui}
\title{Output Shiny UI}
\usage{
simple_ui(id, output, type = "ui", ...)
}
\arguments{
\item{id}{the id of the module. Can be a vector of ids in case of sub module.}

\item{output}{the name of the output}

\item{type}{the type of output: ui (default), plot, table}

\item{...}{other arguments to pass to the *Output function}
}
\value{
the ui object
}
\description{
Basic wrappers around *Output functions
}
\examples{
simple_ui(id = "my_module", output = "my_output")
}
