% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.R
\name{slurm}
\alias{slurm}
\title{Prepare the R script and sh file to submit to HPC managed with SLURM}
\usage{
slurm(
  r.file = "job.r",
  sh.file = paste0(basename(r.file), ".sh"),
  job.name = as.character(r.file),
  time = "1:00:00",
  ntask = 1,
  partition = "fuchs",
  nodes = 1,
  home = "/home/fuchs/fias/knguyen/",
  conda.env = "kinh",
  working.dir = ".",
  submit = FALSE,
  monitor = FALSE,
  user = "knguyen",
  iteration = 1
)
}
\description{
adjust SLURM and conda configuration to your case
}
