% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{source_code}
\alias{source_code}
\title{Source Code}
\usage{
source_code(path, verbose = FALSE)
}
\arguments{
\item{path}{a character vector of the path to explore}

\item{verbose}{a logical value (default = FALSE). If TRUE, then details
about sourced files will be sent to the console}
}
\value{
an integer, how many functions have been attached to the environment
}
\description{
Source all files from a given path.
}
\details{
Given path is explored using list.files with "*.R" pattern, and recursive = TRUE
All listed files are loaded using source()
}
\examples{
\dontrun{
source_code("./R")
}
}
