% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{tabstat}
\alias{tabstat}
\title{format cross-table for publishing}
\usage{
tabstat(.data, ...)
}
\arguments{
\item{row}{row variable}

\item{col}{column (outcome) variable}
}
\description{
format cross-table for publishing, only Fisher/chiquare here
}
