% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{tmb_ICs}
\alias{tmb_ICs}
\title{Calculate Information criteria for TMB model}
\usage{
tmb_ICs(obj, n_post = 1000, pointwise = "pwdens", looic = FALSE)
}
\arguments{
\item{obj}{TMB object}

\item{n_post}{Number of posterior samples}

\item{pointwise}{Name of pointwise predictive density from your model}

\item{looic}{Report leave one out IC from `loo` package?}
}
\description{
Calculate Information criteria for TMB model
}
