% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{tmb_ICs}
\alias{tmb_ICs}
\title{Calculate Information criteria for TMB model}
\usage{
tmb_ICs(post_sample, pointwise = "pwdens", islog = TRUE, looic = FALSE)
}
\arguments{
\item{post_sample}{posterior samples (results from ktools::sample_tmb)}

\item{pointwise}{character name of pointwise predictive density reported from your model}

\item{islog}{Is the pointwise density or log density?}

\item{looic}{Report leave one out IC from `loo` package?}
}
\value{
WAIC-1 and WAIC-2, DIC, and LOO when requested
}
\description{
Note that the model need to report *point-wise density* of the likelihood to
calculate the ICs
}
