% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_level.R
\name{trace_level}
\alias{trace_level}
\title{Set Trace Level}
\usage{
trace_level(level = NULL)
}
\arguments{
\item{level}{an integer to indicate the level of trace printed by
\link[ktools]{catl} function, 0 to set the trace OFF or NULL to get
the current trace level.}
}
\description{
Set Trace Level
}
\examples{
trace_level()
trace_level(2)
}
