% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_r_environ.R
\name{update_r_environ}
\alias{update_r_environ}
\title{Update Renviron File}
\usage{
update_r_environ(path = getwd(), key = "DEBUG", value = "TRUE")
}
\arguments{
\item{path}{the path of the .Renviron file (default is working directory)}

\item{key}{the name of the variable (default = DEBUG)}

\item{value}{the value to be assigned to the variable (default = "TRUE")}
}
\value{
TRUE if success, else FALSE
}
\description{
The function adds a variable to the project .Renviron file.
}
\examples{
\dontrun{
update_r_environ(key = "MY_VAR", value = "my_value")
}
}
