% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_plot_theme.R
\name{use_plot_theme}
\alias{use_plot_theme}
\title{Use Plot Theme Function}
\usage{
use_plot_theme(path = getwd())
}
\arguments{
\item{path}{where to copy the template}
}
\description{
Helper function to create a theme function that can be used across
the project.
}
\examples{
\dontrun{
use_plot_theme()
}
}
