% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_environ.R
\name{use_r_environ}
\alias{use_r_environ}
\title{Create .Renviron File}
\usage{
use_r_environ(path = getwd())
}
\arguments{
\item{path}{the path where to create the file (default set to working directory)}
}
\value{
the output of the file.create call
}
\description{
Create .Renviron File
}
\examples{
\dontrun{
use_r_environ()
}
}
