% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_shiny.R
\name{use_shiny}
\alias{use_shiny}
\title{Setup Shiny App}
\usage{
use_shiny(path = getwd(), app_dir = "shinyapp", data_dir = NULL, module = F)
}
\arguments{
\item{path}{the path where to setup the Shiny app (default = working directory)}

\item{app_dir}{the name of the app folder (default = "shinyapp")}

\item{data_dir}{the path of the data (if any)}

\item{module}{if a Shiny module should be created or not (default: FALSE)}
}
\value{
NULL (invisibly)
}
\description{
The function is used to setup a Shiny App with optional module based on templates.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
When data_dir is null, DATA_HOME environment variable will not be set.
}
\examples{
\dontrun{
use_shiny()
}
}
