% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_shiny.R
\name{use_shiny}
\alias{use_shiny}
\title{Shiny App Setup}
\usage{
use_shiny(
  path = getwd(),
  app_dir = "app",
  data_dir = file.path(path, "data"),
  module = F
)
}
\arguments{
\item{path}{the path where to setup the Shiny app (default: working directory)}

\item{app_dir}{the name of the app folder (default: "app")}

\item{data_dir}{the path of the data folder (default: "data" folder inside path)}

\item{module}{if a Shiny module should be created or not (default: FALSE)}
}
\description{
Shiny App Setup
}
\examples{
\dontrun{
use_shiny()
}
}
