% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid.R
\name{uuid}
\alias{uuid}
\title{Generate uuid}
\usage{
uuid()
}
\value{
a numeric id
}
\description{
Wrapper around getTimestamp()
}
\details{
By default \code{\link[=getTimestamp]{getTimestamp()}} gives a value up to the millisecond.

In some use cases, for example when a sequence of ids need to be generated
grammatically, multiple identical values will be generated within the same
millisecond.

This wrapper calls \code{\link[=getTimestamp]{getTimestamp()}} with \code{k = 1000000} to get a value up to
the microsecond. This is usually enough the get unique value for example
when it's called from lapply.
}
\examples{
uuid()

# showing what the function is trying to solve
any(duplicated(replicate(100, getTimestamp())))
any(duplicated(replicate(100, uuid())))
}
