% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_js.R
\name{with_js}
\alias{with_js}
\title{Manage Package JavaScript Dependency}
\usage{
with_js(package, src, script, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{package}{the name of the package}

\item{src}{where to find the script (see htmltools::htmlDependency() for details)}

\item{script}{the script to include (relative to src)}

\item{session}{the session object of the Shiny server (or shiny::getDefaultReactiveDomain())}
}
\description{
Checks whether a specific dependency is already inserted in the UI with this
function, and perform the insert if necessary.
}
\details{
When \code{session} is NULL, the function will return without doing anything.
This is useful when the function is called from ui side. The JavaScript
dependency will be attached next time the function is called from server side.
This is because the session$userData object cannot be updated from ui side.
}
\examples{
\dontrun{
with_js(package = "ktag", src = "assets", script = "js/ktag.js",
session = shiny::getDefaultReactiveDomain())
}
}
